Remade DeleteMapping.

+++ src/main/java/com/example/serverfortest/repos/TestRepo.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TestRepo extends PagingAndSortingRepository<Test, Long> {
-    Iterable<Test> findById(int i);
+    Iterable<Test> findById(long i);
     List<Test> findByDateLike(String date);
     List<Test> findByTypeLike(String type);
     void deleteByDate(String date);
